Add generic caching class for entire project
It is designed to handle two types of caching, fast and slow.
The fast is dedicated to time based Memcache storage.
The slow is dedicated to the filesystem indefinately.
Fast is to avoid costly/time-consuming external API calls.
Slow is to catch external API failures and be a fallback when all else
fails.