Return the original profile icon URL which should always work
I find this a little difficult to explain, but I will try anyway. The DNS record for planetred might not be the same in directory.unl.edu and gravatar. Because DNS can take some time to propagate, directory will get the new ip address before gravatar. By sending the effective URL to gravatar, there is a potential that the URL might not exist because gravatar is still referring to the old server. By sending the `$profileIconUrl` to gravatar, the request is bound to work regardless of gravatar's dns cache.

The same goes with clients. Client computers might not have the same DNS record as directory.unl.edu and therefor URLs might not resolve correctly.