Remove point of failure from getCategory
Given that getCategoryForAll now handles basically all
caching of this data, we don't need to maintain another call
to the Digital Measures library for individual records. A
new record may not show up for up to 15 minutes in Directory
after being created in Digital Measures, but this is acceptable
as it is the cache TTL for edit requests as well.