fix the qr code link
In some cases, the modal would load with an error because it tried to load based on the target element. That is fine as long as the target element is the `<a href>`, but that isn't the case here. So when the container is clicked, grab the closest `a` element.