Improve filter skipnav logic
This was just relying on the browser to handle a hash change, but that method also had the unfortunate side effect of removing the search results (because the results are also dependant on the url hash). This disables the browser default behavior and instead implments the functionality at the JS layer.