Allow Oracle connection to be turned off
Currently the code makes many calls to Oracle, and if the database
is not found, the timeout on each of these is pretty long. I couldn't
find a good way to turn this down in the codebase. So, for the
All Campus outage this Sunday, I'm implementing this change which will
allow me to turn off this connection temporarily.

Closes #93