Fix oracle performance issues
The query was actually being executed for every record in a search result. So if there were 50 results, there would be 50 more database queries, which slowed the execution to over 30 seconds at times.

To fix this, I moved the logic to be applied to the entire set of search results rather than one record at a time. This has the advantage of keeping the mail attribute correct in the search result JSON output.